fix(mem-cache): resolve macOS clang compilation warnings and method hiding
- Remove duplicate SuperBlk::setBlkSize declaration/definition to prevent method hiding of SectorBlk::setBlkSize.
- Fix sign comparison warning in SectorTags::checkInvariants loop (int vs size_t).
- Fix DPRINTF format specifier for uint64_t offset in CompressedTags::findVictim.

+++ src/mem/cache/tags/compressed_tags.cc
@@ -185,8 +185,8 @@ CompressedTags::findVictim(const CacheBlk::KeyType& key,
         assert(!victim->isValid());
 
         // Print all co-allocated blocks
-        DPRINTF(CacheComp, "Co-Allocation: offset %d of %s\n", offset,
-                victim_superblock->print());
+        DPRINTF(CacheComp, "Co-Allocation: offset %llu of %s\n",
+                (unsigned long long)offset, victim_superblock->print());
     } else {
         victim = victim_superblock->blks[0];
     }
